refactor: catalina/controller 패키지 분리
