+++ tomcat/src/main/java/org/apache/catalina/contoller/HttpMethodCommand.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.catalina.contoller;
 
 import org.apache.coyote.http11.request.Http11Request;
 import org.apache.coyote.http11.response.Http11Response;
